pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "chromadb>=1.4.1",
    "deepagents>=0.3.12",
    "docx2txt>=0.9",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.115.0",
    "langchain>=1.2.9",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ibm>=1.0.3",
    "langchain-openai>=1.1.7",
    "langchain-tavily>=0.2.17",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.9",
    "schedule>=1.2.2",
    "tavily>=1.1.0",
    "uvicorn>=0.30.0",
]

[tool.setuptools.packages.find]
where = ["src"]
